refactor indexation and common elastic
